requirements.txt
# Trade Bot - AI Agent数字货币交易助手
# Python依赖包

# HTTP客户端
httpx>=0.27.0
aiohttp>=3.9.0

# 数据处理
pandas>=2.0.0
numpy>=2.0.0

# 技术指标
pandas-ta>=0.3.14

# K线图绘制
# 注意：mplfinance 官方尚未发布支持 Python 3.13 的稳定版
# 使用预发布版本或从源码安装
mplfinance>=0.12.10b0
matplotlib>=3.9.0

# 异步任务调度
apscheduler>=3.10.0

# 环境变量管理
python-dotenv>=1.0.0

# 数据验证
pydantic>=2.0.0

# OpenAI SDK
openai>=1.0.0

# LangChain - LLM框架
langchain-core>=0.3.0
langchain-openai>=0.2.0

# LangChain 可选提供商（按需安装）
# Anthropic Claude 支持
langchain-anthropic>=0.3.0

# Google Gemini 支持
langchain-google-genai>=2.0.0

# Web框架
fastapi>=0.115.0
uvicorn>=0.30.0
